Changed filter treshold type to double
git-svn-id: https://ipesvn.fzi.de/svn-open/icep/semantic-epa-parent@1268 aa18e314-994e-4353-9a36-64d395711e41

+++ semantic-epa-esper/src/main/java/de/fzi/cep/sepa/esper/filter/numerical/NumericalFilterController.java
@@ -113,7 +113,7 @@ public class NumericalFilterController extends EpDeclarer<NumericalFilterParamet
 		logger.info("Text Property: " +filterProperty);
 	
 		String topicPrefix = "topic://";
-		NumericalFilterParameter staticParam = new NumericalFilterParameter(sepa, Integer.parseInt(threshold), NumericalOperator.valueOf(operation), filterProperty);
+		NumericalFilterParameter staticParam = new NumericalFilterParameter(sepa, Double.parseDouble(threshold), NumericalOperator.valueOf(operation), filterProperty);
 		
 		try {
 			invokeEPRuntime(staticParam, NumericalFilter::new, sepa);
